refactor(feature/entities): 한 폴더 상위로 이동
FSD rules
