+++ src/entities/news/model/index.ts
